requirements.txt
pandas
numpy
yfinance
openai
alpaca-py
pyyaml
lxml
streamlit
